Move Elasticsearch module into Spring-Data-Elasticsearch
