--- elasticsearch/README.md
@@ -1,6 +1,0 @@
-=========
-
-## ElasticSearch
-
-### Relevant Articles: 
-- [A Guide to ElasticSearch](http://www.baeldung.com/????????)
